+++ docs/changes.rst
@@ -8,6 +8,15 @@ This change log is used to track all major changes to Mopidy.
 v0.11.0 (in development)
 ========================
 
+**Settings**
+
+- The settings validator now complains if a setting which expects a tuple of
+  values (e.g. :attr:`mopidy.settings.BACKENDS`,
+  :attr:`mopidy.settings.FRONTENDS`) has a non-iterable value. This typically
+  happens because the setting value contains a single value and one has
+  forgotten to add a comma after the string, making the value a tuple. (Fixes:
+  :issue:`278`)
+
 **Spotify backend**
 
 - Add :attr:`mopidy.settings.SPOTIFY_TIMEOUT` setting which allows you to
@@ -57,6 +66,9 @@ v0.11.0 (in development)
   - Add empty stubs for channel commands for client to client communication.
 
 - Add support for search by date.
+
+- Make ``seek`` and ``seekid`` not restart the current track before seeking in
+  it.
 
 - Include fake tracks representing albums and artists in the search results.
   When these are added to the tracklist, they expand to either all tracks in